refactor: DashboardPage의 상태를 나타내는 부분 컴포넌트화
