+++ frontend/src/pages/DashboardPage.tsx
@@ -2,9 +2,9 @@ import { css } from '@style/css';
 
 import { useParams } from 'react-router-dom';
 
-import { HStack, Text } from '@/components/Common';
 import { SocketProvider } from '@/components/Common/Socket/SocketProvider';
 import DashboardLoading from '@/components/Dashboard/DashboardLoading';
+import DashboardStatus from '@/components/Dashboard/Dashboardstatus';
 import DashboardTable from '@/components/Dashboard/DashboardTable';
 import Header from '@/components/Header';
 import { PageLayout } from '@/components/Layout/PageLayout';
@@ -13,8 +13,8 @@ import { useCompetition } from '@/hooks/competition';
 export default function DashboardPage() {
   const { id } = useParams<{ id: string }>();
   const competitionId: number = id ? parseInt(id, 10) : -1;
-  const { competition } = useCompetition(competitionId);
 
+  const { competition } = useCompetition(competitionId);
   const { startsAt, endsAt } = competition;
   const currentTime = new Date();
 
@@ -45,14 +45,10 @@ export default function DashboardPage() {
         token={localStorage.getItem('accessToken') ?? ''}
         namespace={'dashboards'}
       >
-        <HStack className={textContainerStyle}>
-          <Text type="display" size="lg" className={competitionNameStyle}>
-            {competition.name}
-          </Text>
-          <Text type="title" size="lg">
-            {competitionStatusText}
-          </Text>
-        </HStack>
+        <DashboardStatus
+          competitionName={competition.name}
+          competitionStatusText={competitionStatusText}
+        />
         <DashboardTable useWebsocket={useWebSocket} competitionId={competitionId} />
       </SocketProvider>
     </PageLayout>
@@ -61,18 +57,4 @@ export default function DashboardPage() {
 
 const pageLayoutStyle = css({
   minHeight: '100vh',
-});
-
-const textContainerStyle = css({
-  textAlign: 'center',
-  gap: '10px',
-  marginTop: '56px',
-  marginBottom: '47px',
-});
-
-const competitionNameStyle = css({
-  display: 'flex',
-  height: '116px',
-  flexDirection: 'column',
-  justifyContent: 'center',
 });
